check if key exists

+++ templates/ralfdocs-article-meta.php
@@ -19,19 +19,21 @@ if(!defined('ABSPATH')){ exit; }
       echo '<a href="' . esc_url(get_permalink()) . '" class="meta-btn btn-activities hidden-print" style="background-color:' . esc_html(get_field('activities_color', 'option')) . ';">' . esc_html(sprintf(__('Activities (%d)', 'ralfdocs'), $article_meta['related_activities_count'])) . '</a>';
     }
 
-    if($article_meta['related_impacts_count'] > 0){
+    if(array_key_exists('related_impacts_count', $article_meta) && $article_meta['related_impacts_count'] > 0){
       echo '<a href="' . esc_url(get_permalink()) . '" class="meta-btn btn-impacts hidden-print" style="background-color:' . esc_html(get_field('impacts_color', 'option')) . ';">' . esc_html(sprintf(__('Impacts (%d)', 'ralfdocs'), $article_meta['related_impacts_count'])) . '</a>';
     }
 
-    if($article_meta['resource_types']){
+    if(array_key_exists('resource_types', $article_meta) && $article_meta['resource_types']){
       foreach($article_meta['resource_types'] as $resource_type){
         echo '<a href="' . esc_url($resource_type['link']) . '" class="meta-btn btn-sector hidden-print" style="background-color:' . esc_html($resource_type['color']) . ';">' . esc_html($resource_type['name']) . '</a>';
       }
     }
 
-    $original_resource_url = get_field('original_resource_url', $resource_id);
-    if($original_resource_url){
-      echo '<a href="' . esc_url($original_resource_url) . '" class="meta-btn btn-sector resource-article-type" target="_blank">' . esc_html__('Source', 'ralfdocs') . '</a>';
+    if(isset($resource_id)){
+      $original_resource_url = get_field('original_resource_url', $resource_id);
+      if($original_resource_url){
+        echo '<a href="' . esc_url($original_resource_url) . '" class="meta-btn btn-sector resource-article-type" target="_blank">' . esc_html__('Source', 'ralfdocs') . '</a>';
+      }
     }
   ?>
 </div>
